feat: create componen ModalContent
