+++ src/modules/DictionaryPage/components/StatisticModal/components/ModalContent/index.ts
@@ -1,0 +1,1 @@
+export { ModalContent } from './ModalContent';
