+++ src/modules/DictionaryPage/components/StatisticModal/components/index.ts
@@ -1,0 +1,1 @@
+export { ModalContent } from './ModalContent';
